Moved NewPunch into checklist folder, as that's where it's used
