+++ src/pages/Checklist/ChecklistPage.tsx
@@ -5,7 +5,7 @@ import withAccessControl from '../../services/withAccessControl';
 import useCommonHooks from '../../utils/useCommonHooks';
 import { Route, Switch } from 'react-router-dom';
 import ChecklistWrapper from './ChecklistWrapper';
-import NewPunch, { BottomSpacer } from '../Punch/NewPunch/NewPunch';
+import NewPunch, { BottomSpacer } from './NewPunch/NewPunch';
 import { AsyncStatus } from '../../contexts/CommAppContext';
 import { ChecklistResponse, PunchPreview } from '../../services/apiTypes';
 import { Button } from '@equinor/eds-core-react';
