+++ src/pages/Checklist/NewPunch/NewPunch.tsx
@@ -12,7 +12,7 @@ import { NewPunch as NewPunchType } from '../../../services/apiTypes';
 import NewPunchSuccessPage from './NewPunchSuccessPage';
 import useCommonHooks from '../../../utils/useCommonHooks';
 import styled from 'styled-components';
-import { PunchWrapper } from '../ClearPunch/ClearPunch';
+import { PunchWrapper } from '../../Punch/ClearPunch/ClearPunch';
 import { Button, Scrim } from '@equinor/eds-core-react';
 import {
     AttachmentsWrapper,
